deleted user_input update files + Database
