+++ expenses_management_tool/main_page.py
@@ -2,15 +2,8 @@ import os
 import streamlit as st
 import pandas as pd
 import uuid
-from user_input import store
-from streamlit_function import (
-    enter_data,
-    delete_data,
-    edit_data,
-    view_data,
-    remove_rows,
-)
-
+from useful_functions import store
+import streamlit_function as sf
 # get the right working directory
 root = os.getcwd()
 FILENAME = "expenses_dataset.csv"
@@ -32,7 +25,7 @@ options = ["Enter single entry", "Delete entry", "Edit entry", "View your datafr
 option = st.selectbox("What you want to do", options)
 
 if option == options[0]:
-    df = enter_data()
+    df = sf.enter_data()
     submit = st.button("submit")
     if submit:
         if df["item"][0] != "":  # check input for item
@@ -50,7 +43,7 @@ if option == options[1]:
     df["uuid"] = [uuid.uuid4() for _ in range(len(df.index))]
 
     # create a dataframe, which contain what we want to delete
-    delete_df = delete_data()
+    delete_df = sf.delete_data()
     if delete_df is not None:  # check if delete_df exist
         submit = st.button("Delete")
         if submit:
@@ -59,7 +52,7 @@ if option == options[1]:
             st.write(df)  # old dataframe
 
             # delete row based on unique key
-            df = remove_rows(df, "uuid", delete_df["uuid"])
+            df = sf.remove_rows(df, "uuid", delete_df["uuid"])
             df = df.drop("uuid", axis=1)
             df.to_csv(FILENAME, index=False)  # save to the csv file
             st.write("Your new dataframe!")
@@ -68,7 +61,7 @@ if option == options[1]:
 
 
 if option == options[2]:
-    edit_data()
+    sf.edit_data()
 
 if option == options[3]:
-    view_data()
+    sf.view_data()
