+++ expenses_management_tool/streamlit_function.py
@@ -1,4 +1,3 @@
-import os
 import streamlit as st
 import pandas as pd
